Move all of the utilities into this file
