+++ soxs/tests/test_power_law.py
@@ -9,7 +9,7 @@ from soxs.instrument_registry import \
     get_instrument_from_registry
 from soxs.instrument import instrument_simulator, \
     RedistributionMatrixFile, AuxiliaryResponseFile
-from soxs.tests.utils import write_spectrum, get_wabs_absorb, \
+from soxs.utils import write_spectrum, get_wabs_absorb, \
     convert_rmf
 from sherpa.astro.ui import load_user_model, add_user_pars, \
     load_pha, ignore, fit, set_model, set_stat, set_method, \
